% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beautify.R
\name{beautify}
\alias{beautify}
\alias{beautify_css}
\alias{beautify_html}
\alias{beautify_js}
\title{Beautify html/css/js}
\usage{
beautify_js(text, indent_size = 4, ...)

beautify_html(text, indent_size = 4, ...)

beautify_css(text, indent_size = 4, ...)
}
\arguments{
\item{text}{string with html, css or javascript code.}

\item{indent_size}{number of spaces to indent.}

\item{...}{additional options, see \href{https://www.npmjs.com/package/js-beautify#optionsy}{npm docs}.}
}
\description{
A set of functions to fix indentation and linebreaks within js/html/css code.
The \code{\link{uglify}} function provides a more powerful optimizer and
reformatter for JavaScript.
}
\examples{
test <- "function test(x,y){x=x||1;y=y||1;return x+y;}"
cat(beautify_js(test))
cat(uglify_reformat(test, beautify = TRUE))
}
\references{
Documentation of js-beautify on npm: \url{https://www.npmjs.com/package/js-beautify}
}

