% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee.R
\name{coffee_compile}
\alias{coffee_compile}
\alias{coffee}
\title{Coffee Script}
\usage{
coffee_compile(code, ...)
}
\arguments{
\item{code}{a string with JavaScript code}

\item{...}{additional options passed to the compiler}
}
\description{
Compiles coffee script into JavaScript.
}
\examples{
# Hello world
coffee_compile("square = (x) -> x * x")
coffee_compile("square = (x) -> x * x", bare = TRUE)

# Simple script
demo <- readLines(system.file("example/demo.coffee", package = "js"))
js <- coffee_compile(demo)
cat(js)
cat(uglify_optimize(js))
}
