% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is_js_function.R
\name{is_js_function}
\alias{is_js_function}
\title{Test if code evaluates to a function}
\usage{
is_js_function(text)
}
\arguments{
\item{text}{JavaScript code}
}
\description{
Simple JavaScript function to test if a piece of code evaluates to a function.
}

