% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jade.R
\name{jade}
\alias{jade}
\title{Render Jade Template}
\usage{
jade(text, pretty = FALSE, locals = list())
}
\arguments{
\item{text}{string with jade template.}

\item{pretty}{indent html output.}

\item{locals}{optional list of variables used in the template.}
}
\description{
Jade is a high performance template engine heavily influenced by Haml.
}
\examples{
test.jade <- system.file("examples/test.jade", package = "js")
html <- jade(readLines(test.jade), pretty = TRUE)
cat(html)
}
\references{
Jade documentation: \url{http://jade-lang.com}
}

