% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_eval.R
\name{js_eval}
\alias{js_eval}
\title{Evaluate JavaScript}
\usage{
js_eval(text)
}
\arguments{
\item{text}{JavaScript code}
}
\description{
Evaluate a piece of JavaScript code in a disposable context.
}
\examples{
# Stateless evaluation
js_eval("(function() {return 'foo'})()")

# Use V8 for stateful evaluation
ct <- V8::new_context()
ct$eval("var foo = 123")
ct$get("foo")
}
