% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/js_typeof.R
\name{js_typeof}
\alias{js_typeof}
\title{Get the type of a JavaScript object}
\usage{
js_typeof(text)
}
\arguments{
\item{text}{JavaScript code}
}
\description{
JavaScript wrapper to \code{typeof} to test if a piece of JavaScript code is
syntactically valid, and the type of object it evaluates to. Useful to
verify that a piece of JavaScript code contains a proper function/object.
}
\examples{
js_typeof("function(x){return x+1}")
js_typeof("(function() {return 'foo'})()")
js_typeof("{foo : 123, bar : true}")
}

