% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/minimist.R
\name{minimist}
\alias{minimist}
\title{Parse argument options}
\usage{
minimist(args = commandArgs())
}
\arguments{
\item{args}{a character vector with arguments}
}
\description{
This module is the guts of optimist's argument parser without all the fanciful
decoration.
}
\details{
Parses a vector of strings (e.g. from \code{\link{commandArgs}}) into the proper key/value list.
}
\examples{
minimist()
minimist(c("-a", "beep", "-b", "boop"))
args = c("-x", "3", "-y", "4", "-n5", "-abc", "--beep=boop", "foo", "bar", "baz")
minimist(args)
}

