% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_js}
\alias{validate_js}
\title{Validate JavaScript}
\usage{
validate_js(text)
}
\arguments{
\item{text}{character vector with JavaScript}
}
\description{
Simple wrapper for \code{ct$validate} in \code{\link{V8}}.
}
\examples{
validate_js("function foo(x){2*x}") #TRUE
validate_js("foo = function(x){2*x}") #TRUE

# Anonymous functions in global scope are invalid
validate_js("function(x){2*x}") #FALSE

# But as argument is OK
validate_js("test(function(x){2*x})") #TRUE
}

